[paraformer] separate cif and paraformer (#1795)
