--- examples/aishell/cif/conf/train_paraformer.yaml
@@ -1,88 +1,0 @@
-# network architecture
-# encoder related
-encoder: conformer
-encoder_conf:
-    output_size: 256    # dimension of attention
-    attention_heads: 4
-    linear_units: 2048  # the number of units of position-wise feed forward
-    num_blocks: 12      # the number of encoder blocks
-    dropout_rate: 0.1
-    positional_dropout_rate: 0.1
-    attention_dropout_rate: 0.0
-    input_layer: conv2d # encoder input type, you can chose conv2d, conv2d6 and conv2d8
-    normalize_before: true
-    cnn_module_kernel: 15
-    use_cnn_module: True
-    activation_type: 'swish'
-    pos_enc_layer_type: 'rel_pos'
-    selfattention_layer_type: 'rel_selfattn'
-
-# decoder related
-decoder: transformer
-decoder_conf:
-    attention_heads: 4
-    linear_units: 2048
-    num_blocks: 6
-    dropout_rate: 0.1
-    positional_dropout_rate: 0.1
-    self_attention_dropout_rate: 0.0
-    src_attention_dropout_rate: 0.0
-    input_layer: 'none'
-    src_attention: true
-
-cif_predictor: predictor_v1
-cif_predictor_conf:
-    idim: 256
-    threshold: 1.0
-    l_order: 1
-    r_order: 1
-    tail_threshold: 0.45
-
-# hybrid CTC/attention
-model_conf:
-    ctc_weight: 0.3
-    lsm_weight: 0.1     # label smoothing option
-    length_normalized_loss: false
-    predictor_weight: 1.0
-
-dataset_conf:
-    filter_conf:
-        max_length: 40960
-        min_length: 0
-        token_max_length: 200
-        token_min_length: 1
-    resample_conf:
-        resample_rate: 16000
-    speed_perturb: true
-    fbank_conf:
-        num_mel_bins: 80
-        frame_shift: 10
-        frame_length: 25
-        dither: 0.1
-    spec_aug: true
-    spec_aug_conf:
-        num_t_mask: 2
-        num_f_mask: 2
-        max_t: 50
-        max_f: 10
-    shuffle: true
-    shuffle_conf:
-        shuffle_size: 1500
-    sort: true
-    sort_conf:
-        sort_size: 500  # sort_size should be less than shuffle_size
-    batch_conf:
-        batch_type: 'static' # static or dynamic
-        batch_size: 16
-
-grad_clip: 5
-accum_grad: 4
-max_epoch: 240
-log_interval: 100
-
-optim: adam
-optim_conf:
-    lr: 0.002
-scheduler: warmuplr     # pytorch v1.1.0+ required
-scheduler_conf:
-    warmup_steps: 25000
