+++ examples/aishell/paraformer/README.md
@@ -1,9 +1,7 @@
 # Performance Record
 
-## Standard CIF Conformer Result(TODO)
 
-
-## Paraformer like CIF Conformer Result
+## Paraformer, Conformer Result
 
 * Feature info: using fbank feature, dither, cmvn, online speed perturb
 * Training info: lr 0.002, batch size 16, 4 gpu, acc_grad 4, 240 epochs, dither 0.1
@@ -16,7 +14,7 @@
 | cif greedy search      | 4.41 | 4.92 |
 | cif beam search        | 4.35 | 4.86 |
 
-## Conformer CIF DecoderSANM Result(Deprecated)
+## Paraformer, Conformer DecoderSANM Result(Deprecated)
 
 * Feature info: using fbank feature, dither, cmvn, online speed perturb
 * Training info: lr 0.002, batch size 16, 4 gpu, acc_grad 4, 240 epochs, dither 0.1
