+++ examples/aishell/paraformer/path.sh
